lib/external/fexec.c: always exec in chroot with uid=0
When installing into a rootdir != '/', INSTALL/REMOVE scripts are
currently only executed inside a chroot if uid=0 _and_ the target
rootdir contains an executable bin/sh. Since INSTALL scripts are
executed twice on installation ("pre" and "post" action), an INSTALL
script ran with uid=0 that does the equivalent of `rm -rf /` will wipe
rootdir on the "pre" stage and wipe the real '/' on the "post" stage
(/bin/sh no longer exists -> no chroot).

This removes the check for /bin/sh resulting in always running the
script in a chroot. At least in the uid=0 case.